+++ app/graphql/resolvers.spec.ts
@@ -1,127 +1,178 @@
-import { IncomingMessage } from "http";
+import { RDFCubeViewQueryMock } from "@/test/cube-view-query-mock";
+RDFCubeViewQueryMock;
 
-import { GraphQLResolveInfo } from "graphql";
+// eslint-disable-next-line import/order
+import { NamedNode } from "rdf-js";
 
-import { createSource as createSource_ } from "@/rdf/create-source";
-import { ExtendedCube as ExtendedCube_ } from "@/rdf/extended-cube";
-import { getCubeObservations as getCubeObservations_ } from "@/rdf/queries";
-import { unversionObservation as unversionObservation_ } from "@/rdf/query-dimension-values";
+import { SingleFilters } from "@/config-types";
+import * as ns from "@/rdf/namespace";
+import { getCubeDimensions } from "@/rdf/queries";
+import { getQuery, getQueryFilters } from "@/rdf/query-possible-filters";
+import mockChartConfig from "@/test/__fixtures/config/prod/column-traffic-pollution.json";
 
-import { createContext } from "./context";
-import { Query } from "./resolvers";
-
-const getCubeObservations = getCubeObservations_ as unknown as jest.Mock<
-  typeof getCubeObservations_
->;
-const createSource = createSource_ as unknown as jest.Mock<
-  typeof createSource_
->;
-
-const ExtendedCube = ExtendedCube_ as unknown as jest.Mock<
-  typeof ExtendedCube_
->;
-
-const unversionObservation = unversionObservation_ as unknown as jest.Mock<
-  typeof unversionObservation_
->;
-
-jest.mock("../rdf/query-dimension-values", () => ({
-  unversionObservation: jest.fn(),
-}));
-jest.mock("../rdf/query-search", () => ({}));
-jest.mock("../rdf/queries", () => ({
-  getCubeObservations: jest.fn(),
-  createCubeDimensionValuesLoader: () => async () => [],
-  getLatestCube: () => ({
-    fetchShape: () => ({}),
-  }),
-}));
-jest.mock("../rdf/create-source", () => ({
-  createSource: jest.fn(),
-}));
-jest.mock("../rdf/query-hierarchies", () => ({}));
-jest.mock("../rdf/parse", () => ({}));
-jest.mock("../rdf/extended-cube", () => ({
-  ExtendedCube: jest.fn(),
-}));
-
-jest.mock("@rdf-esm/data-model", () => ({}));
-jest.mock("@rdf-esm/term-map", () => ({}));
-jest.mock("@rdf-esm/namespace", () => ({}));
-jest.mock("@tpluscode/sparql-builder", () => ({}));
-jest.mock("@tpluscode/rdf-ns-builders", () => ({}));
-jest.mock("@tpluscode/rdf-string", () => ({}));
-
-describe("possible filters", () => {
-  beforeEach(() => {
-    getCubeObservations.mockReset();
-  });
-
-  it("should try to find an observation given possible filters, relaxing fitlers from the bottom", async () => {
-    // @ts-ignore
-    getCubeObservations.mockImplementation(async ({ filters }) => {
-      if (Object.keys(filters).length == 2) {
-        return {
-          query: "",
-          observations: [],
-        };
-      } else {
-        return {
-          query: "",
-          observations: [
-            {
-              "https://fake-dimension-iri-1": 1,
-              "https://fake-dimension-iri-2": 3,
+const getCubeDimensionMock = (iri: string, order: string) => {
+  return {
+    path: {
+      value: iri,
+    },
+    out: (p: NamedNode) => {
+      switch (p.value) {
+        case ns.sh.order.value: {
+          return {
+            term: {
+              value: order,
             },
-          ],
-        };
+          };
+        }
+        case ns.cube`meta/dimensionRelation`.value: {
+          return [];
+        }
+        default: {
+          return {
+            out: () => {
+              return {
+                value: undefined,
+                values: [],
+                term: undefined,
+                terms: [],
+                list() {
+                  return [];
+                },
+              };
+            },
+            value: undefined,
+            values: [],
+            term: undefined,
+            terms: [],
+            list() {
+              return [];
+            },
+          };
+        }
       }
+    },
+  };
+};
+
+describe("DataCubeComponents", () => {
+  const fakeCube = {
+    dimensions: [
+      getCubeDimensionMock("dim1", "10"),
+      getCubeDimensionMock("dim2", "0"),
+    ],
+  } as any;
+  it("should return sorted components", async () => {
+    const dimensions = await getCubeDimensions({
+      cube: fakeCube,
+      locale: "en",
+      sparqlClient: {} as any,
+      cache: undefined,
     });
-
-    unversionObservation.mockImplementation(({ observation }) => observation);
-
-    // @ts-ignore
-    createSource.mockImplementation(() => ({
-      cube: () => ({}),
-    }));
-    // @ts-ignore
-    ExtendedCube.mockImplementation(() => ({
-      fetchCube: () => ({}),
-      fetchShape: () => ({}),
-    }));
-
-    const res = await Query?.possibleFilters?.(
-      {},
-      {
-        iri: "https://fake-iri",
-        sourceType: "sparql",
-        sourceUrl: "https://fake-source.com/query",
-        filters: {
-          "https://fake-dimension-iri-1": { type: "single", value: 1 },
-          "https://fake-dimension-iri-2": { type: "single", value: 2 },
-        },
-      },
-      createContext({ req: { headers: {} } as unknown as IncomingMessage }),
-      {
-        variableValues: {
-          locale: "en",
-          sourceUrl: "https://fake-source.com/query",
-        },
-      } as unknown as GraphQLResolveInfo
-    );
-
-    expect(res).toEqual([
-      {
-        iri: "https://fake-dimension-iri-1",
-        type: "single",
-        value: 1,
-      },
-      {
-        iri: "https://fake-dimension-iri-2",
-        type: "single",
-        value: 3,
-      },
-    ]);
-    expect(getCubeObservations).toHaveBeenCalledTimes(2);
+    expect(dimensions.map((d) => d.data.order)).toEqual([0, 10]);
   });
+});
+
+describe("PossibleFilters", () => {
+  const runTest = (
+    cubeIri: string,
+    filters: SingleFilters,
+    versionedDimensionIris: string[],
+    expectedQuery: string
+  ) => {
+    const queryFilters = getQueryFilters(filters, versionedDimensionIris);
+    const query = getQuery(cubeIri, queryFilters);
+    expect(query).toEqual(expectedQuery);
+  };
+
+  it("should generate a correct SPARQL query without ordering", () =>
+    runTest(
+      "cube",
+      {
+        "dim1/2": {
+          type: "single",
+          value: "val1",
+        },
+        dim2: {
+          type: "single",
+          value: "val2",
+        },
+      },
+      ["dim1/2"],
+      `PREFIX cube: <https://cube.link/>
+PREFIX schema: <http://schema.org/>
+PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
+
+SELECT ?dimension0_v ?dimension1 WHERE {
+  <cube> cube:observationSet/cube:observation ?observation .
+  ?observation <dim1/2> ?dimension0 .
+  ?dimension0 schema:sameAs ?dimension0_v .
+  ?observation <dim2> ?dimension1 .
+  VALUES ?dimension0_v { <val1> }
+  BIND(?dimension1 = <val2> AS ?d1)
+}
+ORDER BY DESC(?d1)
+LIMIT 1`
+    ));
+
+  it("should generate a correct SPARQL query with ordering", () =>
+    runTest(
+      "cube",
+      {
+        "dim1/2": {
+          type: "single",
+          value: "val1",
+        },
+        dim2: {
+          type: "single",
+          value: "val2",
+        },
+        dim3: {
+          type: "single",
+          value: "val3",
+        },
+      },
+      ["dim1/2"],
+      `PREFIX cube: <https://cube.link/>
+PREFIX schema: <http://schema.org/>
+PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
+
+SELECT ?dimension0_v ?dimension1 ?dimension2 WHERE {
+  <cube> cube:observationSet/cube:observation ?observation .
+  ?observation <dim1/2> ?dimension0 .
+  ?dimension0 schema:sameAs ?dimension0_v .
+  ?observation <dim2> ?dimension1 .
+  ?observation <dim3> ?dimension2 .
+  VALUES ?dimension0_v { <val1> }
+  BIND(?dimension1 = <val2> AS ?d1)
+  BIND(?dimension2 = <val3> AS ?d2)
+}
+ORDER BY DESC(?d1) DESC(?d2)
+LIMIT 1`
+    ));
+
+  it("should generate a correct SPARQL query based on real cube", () =>
+    runTest(
+      mockChartConfig.data.dataSet,
+      mockChartConfig.data.chartConfig.filters as SingleFilters,
+      // assumption: the versioned dimension iris are the same as the keys of the filters
+      Object.keys(mockChartConfig.data.chartConfig.filters),
+      `PREFIX cube: <https://cube.link/>
+PREFIX schema: <http://schema.org/>
+PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
+
+SELECT ?dimension0_v ?dimension1_v ?dimension2_v WHERE {
+  <https://environment.ld.admin.ch/foen/ubd003701/1> cube:observationSet/cube:observation ?observation .
+  ?observation <https://environment.ld.admin.ch/foen/ubd003701/beurteilung> ?dimension0 .
+  ?dimension0 schema:sameAs ?dimension0_v .
+  ?observation <https://environment.ld.admin.ch/foen/ubd003701/gemeindetype> ?dimension1 .
+  ?dimension1 schema:sameAs ?dimension1_v .
+  ?observation <https://environment.ld.admin.ch/foen/ubd003701/laermbelasteteeinheit> ?dimension2 .
+  ?dimension2 schema:sameAs ?dimension2_v .
+  VALUES ?dimension0_v { <https://environment.ld.admin.ch/foen/ubd003701/beurteilung/%3EIGWLSV> }
+  BIND(?dimension1_v = <https://environment.ld.admin.ch/foen/ubd003701/gemeindeTyp/CH> AS ?d1)
+  BIND(?dimension2_v = <https://environment.ld.admin.ch/foen/ubd003701/laermbelasteteEinheit/Pers> AS ?d2)
+}
+ORDER BY DESC(?d1) DESC(?d2)
+LIMIT 1`
+    ));
 });