+++ playwright.config.ts
@@ -28,7 +28,7 @@ const config: PlaywrightTestConfig = {
   forbidOnly: !!process.env.CI,
   /* Retry on CI only */
   retries: process.env.CI ? 1 : 0,
-  workers: 1,
+  workers: 4,
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
   reporter: process.env.CI ? "list" : "line",
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */