+++ CHANGELOG.md
@@ -11,6 +11,8 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
 
 - Performance
   - Hierarchies are now fetched in parallel with dimension values, rather than waiting for them to be fetched in advance
+- Maintenance
+  - Increased the number of Playwright workers to 4
 
 # [3.27.0] - 2024-03-12
 
@@ -27,6 +29,8 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
   - Currently selected locale is now persisted when using actions in user profile page (view chart, edit, open, share)
   - Initial zoom map is now correctly (and consistently) applied again
   - Optional filters can be added again
+- Performance
+  - Vastly optimized the `PossibleFilters` query
 
 # [3.26.3] - 2024-03-05